% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{add_buttons_to_table}
\alias{add_buttons_to_table}
\title{Add buttons to data frame}
\usage{
add_buttons_to_table(df, table_id, id_col, label = "Select")
}
\arguments{
\item{df}{Data frame}

\item{table_id}{Unique id for data frame}

\item{id_col}{Unique key column in data frame}

\item{label}{Label to put on button, default is "Select"}
}
\value{
Data frame with buttons added
}
\description{
Add buttons to data frame
}
