% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis.R
\name{analysis_enforcing_chars_ui}
\alias{analysis_enforcing_chars_ui}
\title{UI for Analysis Enforcing Characters}
\usage{
analysis_enforcing_chars_ui(id, id_lists)
}
\arguments{
\item{id}{Id of module}

\item{id_lists}{Id of tree_lists module}
}
\value{
A shiny tagList for the analysis enforcing characters UI
}
\description{
UI for Analysis Enforcing Characters
}
