% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rel_chron.R
\name{analysis_rel_chron_server}
\alias{analysis_rel_chron_server}
\title{Server for Relative Chronology}
\usage{
analysis_rel_chron_server(
  id,
  my_vals,
  cache,
  analysis_root_drop,
  tree_to_listen,
  data_upload_root_drop,
  paup_is_weighted,
  analysis_metrics
)
}
\arguments{
\item{id}{Id of module}

\item{my_vals}{Reactive variables}

\item{cache}{Cache object}

\item{analysis_root_drop}{Root from analysis page}

\item{tree_to_listen}{Radio box selection}

\item{data_upload_root_drop}{Root from data upload page}

\item{paup_is_weighted}{Whether or not PAUP* was run with weight}

\item{analysis_metrics}{Metrics to include in analysis}
}
\value{
Nothing
}
\description{
Server for Relative Chronology
}
