% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rel_chron.R
\name{analysis_rel_chron_ui}
\alias{analysis_rel_chron_ui}
\title{UI for Relative Chronology}
\usage{
analysis_rel_chron_ui(id, id_lists)
}
\arguments{
\item{id}{Id of module}

\item{id_lists}{Id of tree_lists module}
}
\value{
A shiny tagList for the relative chronology UI
}
\description{
UI for Relative Chronology
}
