% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis.R
\name{analysis_server}
\alias{analysis_server}
\title{Server for Analysis}
\usage{
analysis_server(
  id,
  my_vals,
  cache,
  data_upload_root_drop,
  table_button,
  tree_to_listen,
  paup_is_weighted
)
}
\arguments{
\item{id}{Id of module}

\item{my_vals}{Reactive values}

\item{cache}{Cache object}

\item{data_upload_root_drop}{Root from data upload page}

\item{table_button}{Table button reactive value}

\item{tree_to_listen}{Tree to do analysis on}

\item{paup_is_weighted}{Whether or not PAUP* is weighted}
}
\value{
List of reactive values
}
\description{
Server for Analysis
}
