% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tree_lists.R
\name{analysis_tree_lists_server}
\alias{analysis_tree_lists_server}
\title{Server for Tree Lists}
\usage{
analysis_tree_lists_server(
  id,
  my_vals,
  cache,
  data_upload_root_drop,
  paup_is_weighted,
  analysis_metrics
)
}
\arguments{
\item{id}{Id of module}

\item{my_vals}{Reactive variables}

\item{cache}{Cache object}

\item{data_upload_root_drop}{Root from data upload page}

\item{paup_is_weighted}{Whether or not PAUP* is weighted}

\item{analysis_metrics}{List of analysis metrics}
}
\value{
Tree to be rendered
}
\description{
Server for Tree Lists
}
