% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tree_lists.R
\name{analysis_tree_lists_ui}
\alias{analysis_tree_lists_ui}
\title{UI for Tree Lists}
\usage{
analysis_tree_lists_ui(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
Shiny tagList for Tree Lists UI
}
\description{
UI for Tree Lists
}
