% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis.R
\name{analysis_tree_ui}
\alias{analysis_tree_ui}
\title{UI for Analysis Tree Diagram}
\usage{
analysis_tree_ui(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
UI for analysis page
}
\description{
UI for Analysis Tree Diagram
}
