% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_analysis.R
\name{analysis_ui_sidebar}
\alias{analysis_ui_sidebar}
\title{UI Sidebar for Analysis}
\usage{
analysis_ui_sidebar(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
A shiny tagList for the analysis UI sidebar
}
\description{
UI Sidebar for Analysis
}
