% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate tree with states on internal nodes}
\usage{
annotate_tree(tree, char_rep, parsimony_rep, edge_orig)
}
\arguments{
\item{tree}{Tree to annotate}

\item{char_rep}{Character representation with which tree is to be annotated}

\item{parsimony_rep}{Parsimony result from running maximum parsimony on
this tree with this character}

\item{edge_orig}{Edge table of tree with original root; if anc_taxon
was originally set, then should be NULL}
}
\value{
List with annotations by edge id and mapping variable
}
\description{
Annotate tree with states on internal nodes
}
