% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Shiny Server for LinguiPhyR}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input for server}

\item{output}{Output for server}

\item{session}{Session variable}
}
\description{
Shiny Server for LinguiPhyR
}
