% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{cache_wrapper}
\alias{cache_wrapper}
\title{cache wrapper}
\usage{
cache_wrapper(cache, cache_id, code_block)
}
\arguments{
\item{cache}{cache object}

\item{cache_id}{Id of key in cache}

\item{code_block}{Code to execute}
}
\value{
Result of code
}
\description{
cache wrapper
}
