% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{collapse_edge}
\alias{collapse_edge}
\title{Collapse edge on tree}
\usage{
collapse_edge(tree, node1, node2)
}
\arguments{
\item{tree}{Tree on which to collapse edge}

\item{node1}{One endpoint of edge to collapse}

\item{node2}{Other endpoint of edge to collapse}
}
\value{
Tree with collapsed edge
}
\description{
Collapse edge on tree
}
