% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_upload.R
\name{data_upload_server}
\alias{data_upload_server}
\title{Server for Data Upload Module}
\usage{
data_upload_server(id, my_vals)
}
\arguments{
\item{id}{Id of module}

\item{my_vals}{Reactive values}
}
\value{
List with reactive values
}
\description{
Server for Data Upload Module
}
