% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_upload.R
\name{data_upload_ui}
\alias{data_upload_ui}
\title{UI for Data Upload}
\usage{
data_upload_ui(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
UI for Data Upload
}
\description{
UI for Data Upload
}
