% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_upload.R
\name{data_upload_ui_sidebar}
\alias{data_upload_ui_sidebar}
\title{UI Sidebar for Data Upload}
\usage{
data_upload_ui_sidebar(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
A shiny tagList for the data upload UI sidebar
}
\description{
UI Sidebar for Data Upload
}
