% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{find_chars_incompat_any_tree}
\alias{find_chars_incompat_any_tree}
\title{Find characters incompatible on any tree}
\usage{
find_chars_incompat_any_tree(char_reps, parsimony_caches, names = NULL)
}
\arguments{
\item{char_reps}{Character representations}

\item{parsimony_caches}{Parsimony cache on each tree}

\item{names}{Names of trees, if wish to include in character, else NULL}
}
\value{
Data frame containing characters incompatible on any tree
}
\description{
Find characters incompatible on any tree
}
