% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{find_chars_supporting_clade}
\alias{find_chars_supporting_clade}
\title{Find characters in data frame that support clade (tree-agnostic)}
\usage{
find_chars_supporting_clade(clade, df)
}
\arguments{
\item{clade}{Character vector containing languages of clade}

\item{df}{Linguistic data frame}
}
\value{
Data frame containing only characters that support provided clade
}
\description{
Find characters in data frame that support clade (tree-agnostic)
}
