% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{find_enforcing_chars}
\alias{find_enforcing_chars}
\title{Find characters that enforce each edge on phylogenetic tree}
\usage{
find_enforcing_chars(tree, char_reps, parsimony_cache)
}
\arguments{
\item{tree}{Binary phylogenetic tree}

\item{char_reps}{Character representations}

\item{parsimony_cache}{Parsimony cache containing parsimony information on
characters}
}
\value{
List with information about enforcing characters for each edge
}
\description{
A character is said to enforce (or support) an edge in a
  phylogenetic tree if the edge's collapse would increase the parsimony
  score of the character on the tree.
}
