% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{find_incompat_chars}
\alias{find_incompat_chars}
\title{Find incompatible characters on tree}
\usage{
find_incompat_chars(char_reps, parsimony_cache, return_all_chars = FALSE)
}
\arguments{
\item{char_reps}{Character representations}

\item{parsimony_cache}{Parsimony information}

\item{return_all_chars}{Whether or not to return all characters or just the
incompatible ones}
}
\value{
Data frame with incompatible characters
}
\description{
Find incompatible characters on tree
}
