% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{find_uninformative_characters}
\alias{find_uninformative_characters}
\title{Find parsimony-uninformative characters in data frame}
\usage{
find_uninformative_characters(df)
}
\arguments{
\item{df}{Linguistic data frame}
}
\value{
List with row numbers of uninformative characters and number of
  big states per character
}
\description{
Find parsimony-uninformative characters in data frame
}
