% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{generate_error}
\alias{generate_error}
\title{Generate error message in Shiny}
\usage{
generate_error(msg, frame = NULL)
}
\arguments{
\item{msg}{Message}

\item{frame}{Data frame (optional)}
}
\value{
Nothing
}
\description{
Generate error message in Shiny
}
