% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsimony.R
\name{get_paup_output_strings}
\alias{get_paup_output_strings}
\title{Get output strings from PAUP*}
\usage{
get_paup_output_strings()
}
\value{
List with two elements: PAUP* input file and PAUP* output file
}
\description{
Get output strings from PAUP*
}
