% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_tree_to_use}
\alias{get_tree_to_use}
\title{Get tree to use based on radio box}
\usage{
get_tree_to_use(
  tree_to_listen,
  my_vals,
  cache,
  data_upload_root_drop,
  paup_is_weighted,
  analysis_metrics
)
}
\arguments{
\item{tree_to_listen}{Radio box selection}

\item{my_vals}{Reactive variables}

\item{cache}{Cache object}

\item{data_upload_root_drop}{Root from data upload page}

\item{paup_is_weighted}{Whether or not PAUP* was run with weight}

\item{analysis_metrics}{Metrics to include in analysis}
}
\value{
Tree corresponding to selection
}
\description{
Get tree to use based on radio box
}
