% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{get_uninf_char_text}
\alias{get_uninf_char_text}
\title{Get text for parsimony-uninformative characters}
\usage{
get_uninf_char_text(uninf_chars, df)
}
\arguments{
\item{uninf_chars}{Parsimony-uninformative characters}

\item{df}{Data frame with character information}
}
\value{
Text to display regarding parsimony-uninformative characters
}
\description{
Get text for parsimony-uninformative characters
}
