% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pipe.R
\name{\%<+\%}
\alias{\%<+\%}
\title{Ggtree backward operator}
\arguments{
\item{lhs}{A value}

\item{rhs}{A value}
}
\value{
The result of %<+%.
}
\description{
See https://bioc.ism.ac.jp/packages/3.5/bioc/vignettes/ggtree/inst/doc/
treeAnnotation.html#tree-annotation-with-user-specified-annotation
}
\keyword{internal}
