% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{join_enf_chars_with_edge_table}
\alias{join_enf_chars_with_edge_table}
\title{Join enforcing characters table with edge table}
\usage{
join_enf_chars_with_edge_table(enforcing_chars, edge)
}
\arguments{
\item{enforcing_chars}{Enforcing characters table}

\item{edge}{Edge table of tree}
}
\value{
Table with enforcing character count and summed weight by
  edge using parent and child ids from edge table
}
\description{
Join enforcing characters table with edge table
}
