% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_character.R
\name{make_character_representation}
\alias{make_character_representation}
\title{Make character representation}
\usage{
make_character_representation(
  row,
  root = NULL,
  do_ancstate = FALSE,
  neutralize_custom = FALSE,
  overwrite_custom_weight = FALSE,
  symbols = paste0("ABCDEFGHIJKLMNOP", "QRSTUVWXYZabcdef", "ghijklmnopqrstuv", "wxyz")
)
}
\arguments{
\item{row}{Row of linguistic data frame representing a character}

\item{root}{Leaf representing root if available, otherwise NULL}

\item{do_ancstate}{Whether or not to use ancestral state in representation}

\item{neutralize_custom}{Whether or not to set all custom costs to 1}

\item{overwrite_custom_weight}{Set character weight to max cost in custom
weight matrix. If TRUE, neutralize_custom must also be TRUE}

\item{symbols}{Alphabetic symbols in the order with which to replace
state numbers}
}
\value{
List, containing several attributes that represent various aspects
  of the character
}
\description{
Make character representation
}
