% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_character.R
\name{make_character_representations}
\alias{make_character_representations}
\title{Make character representations}
\usage{
make_character_representations(
  root,
  data,
  neutralize_custom,
  overwrite_custom_weight
)
}
\arguments{
\item{root}{Root of tree if available, otherwise NULL}

\item{data}{Data frame}

\item{neutralize_custom}{Whether or not to neutralize custom weights}

\item{overwrite_custom_weight}{Set character weight to max cost in custom
weight matrix. If TRUE, neutralize_custom must also be TRUE}
}
\value{
List of character representations, or data frame containing errors
}
\description{
Make character representations
}
