% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_relative_chronology.R
\name{make_chronology}
\alias{make_chronology}
\title{Make relative chronology for tree}
\usage{
make_chronology(tree, clade, char_reps, parsimony_cache, root)
}
\arguments{
\item{tree}{Tree on which to compute relative chronology}

\item{clade}{Clade for which to compute relative chronology}

\item{char_reps}{Characters to include in the relative chronology}

\item{parsimony_cache}{Parsimony results on each character on this tree}

\item{root}{Root of tree}
}
\value{
Relative chronology data frame
}
\description{
Make relative chronology for tree
}
