% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_relative_chronology.R
\name{make_chronology_for_char}
\alias{make_chronology_for_char}
\title{Make relative chronology for this character}
\usage{
make_chronology_for_char(tree, clade, char_rep, parsimony_rep, root)
}
\arguments{
\item{tree}{Tree on which to compute relative chronology}

\item{clade}{Clade for which to compyute relative chronology}

\item{char_rep}{Character represntation}

\item{parsimony_rep}{Parsimony representation for this character}

\item{root}{Root of the tree}
}
\value{
List with relative chronology information, or NULL if clade is not
  a clade of the tree
}
\description{
Make relative chronology for this character
}
