% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{make_consensus_trees}
\alias{make_consensus_trees}
\title{Return strict and greedy consensus trees}
\usage{
make_consensus_trees(trees)
}
\arguments{
\item{trees}{Trees for which to compute consensus trees}
}
\value{
List with consensus trees and names of those trees
}
\description{
Return strict and greedy consensus trees
}
