% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_character.R
\name{make_cost_matrix_for_parsimony}
\alias{make_cost_matrix_for_parsimony}
\title{Make cost matrix for asr_max_parsimony.}
\usage{
make_cost_matrix_for_parsimony(
  unique_states,
  total_num_states,
  style,
  custom_matrix = NULL
)
}
\arguments{
\item{unique_states}{Unique states vector (must be in order)}

\item{total_num_states}{Total number of states (including ? states)}

\item{style}{One of "standard", "irreversible", or "custom"}

\item{custom_matrix}{Custom matrix if style == "custom" else NULL}
}
\value{
Cost matrix for asr_max_parsimony. Includes rows/columns for all ?
  characters.
}
\description{
Make cost matrix for asr_max_parsimony.
}
