% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_character.R
\name{make_custom_matrix}
\alias{make_custom_matrix}
\title{Make custom transition cost matrix}
\usage{
make_custom_matrix(string, states_char)
}
\arguments{
\item{string}{Custom matrix specified as a string}

\item{states_char}{Unique states vector (must be in order)}
}
\value{
Custom transition cost matrix
}
\description{
Make custom transition cost matrix
}
