% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_datatable}
\alias{make_datatable}
\title{Make DT::datatable from data frame}
\usage{
make_datatable(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
datatable
}
\description{
Make DT::datatable from data frame
}
