% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{make_edges}
\alias{make_edges}
\title{Number edges in tree (invariant to root placement)}
\usage{
make_edges(tree)
}
\arguments{
\item{tree}{Tree For which to produce edge numbers}
}
\value{
List of edges, which is invariant to the root placement
}
\description{
Number edges in such a way that result is invariant to root
  placement. Edge ids are assigned by bipartition, not internal node
  labelings. Note that the internal nodes of tree$edge change depending on
  the root placement, but the edge id's produced are invariant to this.
  Thus, this function is useful for many downstream applications, as one
  can use edge id to reference parts of the tree regardless of root
  location.
}
