% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_character.R
\name{make_irreverseible_matrix}
\alias{make_irreverseible_matrix}
\title{Make cost matrix for irreversible character}
\usage{
make_irreverseible_matrix(num_states)
}
\arguments{
\item{num_states}{Number of states}
}
\value{
Cost matrix
}
\description{
Make cost matrix for irreversible character
}
