% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_trees_io.R
\name{make_tree_names}
\alias{make_tree_names}
\title{Make tree names}
\usage{
make_tree_names(trees, names_to_do)
}
\arguments{
\item{trees}{List of trees}

\item{names_to_do}{Names of score types to include}
}
\value{
Character vector of names corresponding to trees
}
\description{
Make tree names
}
