% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_tree_plot_wrapper}
\alias{make_tree_plot_wrapper}
\title{UI Wrapper for make_treeplot}
\usage{
make_tree_plot_wrapper(
  tree_to_use,
  cache,
  root_drop,
  show_edge_ids,
  show_edge_support,
  show_weighted_edge_support,
  root_on_left,
  height,
  width,
  char_to_put = NULL
)
}
\arguments{
\item{tree_to_use}{Tree to plot}

\item{cache}{Cache object}

\item{root_drop}{Leaf to root tree at}

\item{show_edge_ids}{Whether or not to show edge ids}

\item{show_edge_support}{Whether or not to show edge support}

\item{show_weighted_edge_support}{Whether or not to show weighted
edge support}

\item{root_on_left}{Whether or not root should show on left (i.e.
as unrooted)}

\item{height}{Height of plot}

\item{width}{Width of plot}

\item{char_to_put}{Optionally, a character with which to annotate tree}
}
\value{
Tree plot
}
\description{
UI Wrapper for make_treeplot
}
