% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{make_treeplot}
\alias{make_treeplot}
\title{Make tree plot}
\usage{
make_treeplot(
  orig_tree,
  root,
  do_edge_label,
  node_states = NULL,
  enforcing_chars = NULL,
  show_root_at_left = FALSE,
  message = NULL,
  precomputed_edge = NULL,
  show_edge_support = FALSE,
  show_weighted_edge_support = FALSE,
  height = 400,
  width = 800
)
}
\arguments{
\item{orig_tree}{Tree to display}

\item{root}{Root node to display tree at}

\item{do_edge_label}{Whether or not to label edges}

\item{node_states}{States to display at nodes if available, else NULL}

\item{enforcing_chars}{Characters that enforce edges if available, else NULL}

\item{show_root_at_left}{Whether or not to show root on left}

\item{message}{Message to display on plot}

\item{precomputed_edge}{Edge information}

\item{show_edge_support}{Whether or not to show edge support}

\item{show_weighted_edge_support}{Whether or not to show weighted edge
support}

\item{height}{Height of plot}

\item{width}{Width of plot}
}
\value{
Plotly plot
}
\description{
Make tree plot
}
