% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_paup.R
\name{paup_server}
\alias{paup_server}
\title{Server for PAUP*}
\usage{
paup_server(id, my_vals, cache, data_upload_root_drop)
}
\arguments{
\item{id}{Id of module}

\item{my_vals}{Reactive variables}

\item{cache}{Cache object}

\item{data_upload_root_drop}{Root from data upload page}
}
\value{
List of reactive variables
}
\description{
Server for PAUP*
}
