% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_paup.R
\name{paup_ui_sidebar}
\alias{paup_ui_sidebar}
\title{UI Sidebar for PAUP*}
\usage{
paup_ui_sidebar(id)
}
\arguments{
\item{id}{Id of module}
}
\value{
A shiny tagList for the PAUP* UI sidebar
}
\description{
UI Sidebar for PAUP*
}
