% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocessing the linguistic data}
\usage{
preprocess_data(df, poly_choice)
}
\arguments{
\item{df}{Data frame to preprocess}

\item{poly_choice}{String, how to handle polymorphism}
}
\value{
List containing:
  df_with_poly: Data frame containing the polymorphisms
  df: Preprocessed data frame
  num_poly: Numeric vector containing number of polymorphic states per
    character
  polywid: Numeric vector containing the maximum number of polymorphic
    states per character
}
\description{
Preprocessing the linguistic data
}
