% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{read_and_name_trees_cache}
\alias{read_and_name_trees_cache}
\title{Read and name trees using cache}
\usage{
read_and_name_trees_cache(cache, cache_id, cache_code, names_to_do)
}
\arguments{
\item{cache}{cache object}

\item{cache_id}{Id of key in cache}

\item{cache_code}{Code to execute}

\item{names_to_do}{Names to label the trees}
}
\value{
Trees with names
}
\description{
Read and name trees using cache
}
