% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_trees_io.R
\name{read_and_name_trees_file}
\alias{read_and_name_trees_file}
\title{Read and name trees from file}
\usage{
read_and_name_trees_file(
  file,
  char_reps,
  leafset,
  is_weighted,
  insert_root,
  do_unroot = TRUE
)
}
\arguments{
\item{file}{File path}

\item{char_reps}{Character representations}

\item{leafset}{Leafs to include}

\item{is_weighted}{Whether or not to use weighted parsimony for sorting}

\item{insert_root}{Root leaf to insert}

\item{do_unroot}{Whether or not to unroot trees}
}
\value{
List of trees along with their names to display
}
\description{
Read and name trees from file
}
