% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_trees_io.R
\name{read_and_name_trees_paup}
\alias{read_and_name_trees_paup}
\title{Read and name trees from PAUP*}
\usage{
read_and_name_trees_paup(char_reps, is_weighted, do_unroot, insert_root)
}
\arguments{
\item{char_reps}{Character representations}

\item{is_weighted}{Whether or not to use weighted parsimony for sorting}

\item{do_unroot}{Whether or not to unroot trees}

\item{insert_root}{Root leaf to insert}
}
\value{
List of trees along with their names to display
}
\description{
Read and name trees from PAUP*
}
