% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{read_and_validate_data}
\alias{read_and_validate_data}
\title{Read linguistic data and do basic validation}
\usage{
read_and_validate_data(path)
}
\arguments{
\item{path}{A string pointing to the file path}
}
\value{
A data frame if successful, or a string containing error message
if not
}
\description{
Read linguistic data and do basic validation
}
