% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_trees_io.R
\name{read_trees_from_paup}
\alias{read_trees_from_paup}
\title{Read trees from PAUP* file}
\usage{
read_trees_from_paup(do_unroot, insert_root = NULL)
}
\arguments{
\item{do_unroot}{Whether or not to unroot the trees}

\item{insert_root}{Whether or not to insert root into trees}
}
\value{
List of trees read from PAUP* file
}
\description{
Read trees from PAUP* file
}
