% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{replace_qs_with_nums}
\alias{replace_qs_with_nums}
\title{Replace ?'s with unique numbers in linguistic data frame}
\usage{
replace_qs_with_nums(df)
}
\arguments{
\item{df}{Linguistic data frame}
}
\value{
Linguistic data frame, where all ?'s are replaced with unique
  numbers
}
\description{
Replace ?'s with unique numbers in linguistic data frame
}
