% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ringe_screened_dataset}
\alias{ringe_screened_dataset}
\alias{ringe_dataset}
\title{Screened Indo-European dataset from Ringe et al., 2002}
\format{
A CSV file containing linguistic characters for 24 Indo-European
  languages.
}
\source{
Ringe et al., 2002.
  https://onlinelibrary.wiley.com/doi/abs/10.1111/1467-968X.00091
}
\usage{
ringe_dataset
}
\description{
This is the screened character dataset from Ringe et al., 2002.
}
\seealso{
https://tandy.cs.illinois.edu/histling.html
}
\keyword{datasets}
