% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsimony.R
\name{run_parsimony}
\alias{run_parsimony}
\title{Compute parsimony for character on given tree}
\usage{
run_parsimony(tree, char_rep)
}
\arguments{
\item{tree}{Tree on which to calculate parsimony}

\item{char_rep}{Character representation of character for which to compute
parsimony}
}
\value{
List with parsimony information.
}
\description{
Compute parsimony for character on given tree
}
