% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsimony.R
\name{run_parsimony_on_each_char}
\alias{run_parsimony_on_each_char}
\title{Compute parsimony for each character on given tree}
\usage{
run_parsimony_on_each_char(tree, char_reps)
}
\arguments{
\item{tree}{Tree on which to calculate parsimony}

\item{char_reps}{Character representations for which to compute parsimony}
}
\value{
List of parsimony results on each character on the provided tree
}
\description{
Compute parsimony for each character on given tree
}
