% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{show_loading_for_paup}
\alias{show_loading_for_paup}
\title{Show PAUP* loading message and progress bar}
\usage{
show_loading_for_paup(is_exhaustive)
}
\arguments{
\item{is_exhaustive}{Whether or not PAUP* is being run in exhaustive mode}
}
\value{
Nothing
}
\description{
Show PAUP* loading message and progress bar
}
