% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_trees_io.R
\name{sort_and_name_trees}
\alias{sort_and_name_trees}
\title{Sort and name trees}
\usage{
sort_and_name_trees(trees, char_reps, sort_by = "wmp", names = NULL)
}
\arguments{
\item{trees}{List of phylogenetic trees}

\item{char_reps}{Character representations}

\item{sort_by}{Metric to sort by, either 'wmp' for weighted maximum
parsimony or 'mp' for maximum parsimony}

\item{names}{Names of trees if pre-specified, else NULL}
}
\value{
List of trees with names and scores
}
\description{
Sort and name trees
}
