% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsimony.R
\name{write_nexus}
\alias{write_nexus}
\title{Write NEXUS file for PAUP* to run maximum parsimony}
\usage{
write_nexus(
  char_reps,
  taxa,
  do_weight,
  is_exhaustive,
  keep,
  symbols = paste0("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefg", "hijklmnopqrstuvwxyz"),
  root = NULL
)
}
\arguments{
\item{char_reps}{Character representations}

\item{taxa}{Taxa}

\item{do_weight}{Whether or not to run weighted parsimony}

\item{is_exhaustive}{Whether or not to run exhaustive search}

\item{keep}{How many trees to keep}

\item{symbols}{Symbols to use when writing file}

\item{root}{Root of tree}
}
\value{
String containing text to write as NEXUS file
}
\description{
Write NEXUS file for PAUP* to run maximum parsimony
}
